kinetic-api/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
supabase==2.28.3
python-dotenv==1.2.2
pydantic==2.12.5
PyJWT==2.12.1
httpx==0.28.1
python-dateutil==2.9.0.post0
